+++ src/definition/Definition.js
@@ -405,6 +405,7 @@ export function Definition(context, options) {
 
   function getApi(message) {
     if (execution) return execution.getApi(message);
+    if (!message || !stateMessage) throw new Error('Definition is not running');
     return DefinitionApi(broker, message || stateMessage);
   }
 
@@ -420,7 +421,6 @@ export function Definition(context, options) {
       const resolvedReference = reference.resolve(createMessage({message}));
       messageType = resolvedReference.messageType || messageType;
       messageContent.message = {...message, ...resolvedReference};
-
     }
 
     return getApi().sendApiMessage(messageType, messageContent, {delegate: true});