+++ src/eventDefinitions/MessageEventDefinition.js
@@ -3,7 +3,7 @@ import {brokerSafeId} from '../shared';
 import {cloneContent, shiftParent} from '../messageHelper';
 
 export default function MessageEventDefinition(activity, eventDefinition) {
-  const {id, broker, environment, isThrowing, getActivityById} = activity;
+  const {id, broker, environment, isStart, isThrowing, getActivityById} = activity;
   const {type = 'MessageEventDefinition', behaviour = {}} = eventDefinition;
   const {debug} = environment.Logger(type.toLowerCase());
   const reference = behaviour.messageRef || {name: 'anonymous'};
@@ -11,7 +11,7 @@ export default function MessageEventDefinition(activity, eventDefinition) {
   const messageId = referenceElement ? referenceElement.id : 'anonymous';
   const messageQueueName = `message-${brokerSafeId(id)}-${brokerSafeId(messageId)}-q`;
 
-  if (!isThrowing) setupCatch();
+  if (!isThrowing || isStart) setupCatch();
 
   const source = {
     id,
@@ -36,6 +36,7 @@ export default function MessageEventDefinition(activity, eventDefinition) {
 
     broker.subscribeTmp('api', `activity.#.${executionId}`, onApiMessage, {noAck: true, consumerTag: `_api-${executionId}`, priority: 400});
     if (parentExecutionId) broker.subscribeTmp('api', `activity.#.${parentExecutionId}`, onApiMessage, {noAck: true, consumerTag: `_api-parent-${executionId}`, priority: 400});
+    broker.subscribeTmp('api', '#.signal.*', onCatchMessage, {noAck: true, consumerTag: `_api-delegated-${executionId}`});
 
     debug(`<${executionId} (${id})> expect ${description}`);
 
@@ -48,20 +49,24 @@ export default function MessageEventDefinition(activity, eventDefinition) {
 
     function onCatchMessage(routingKey, message) {
       if (getPropertyValue(message, 'content.message.id') !== referenceMessage.id) return;
-      complete('caught', message.content.message);
+
+      const {type: messageType, correlationId} = message.properties;
+      broker.publish('event', 'activity.consumed', cloneContent(messageContent, {message: {...message.content.message}}), {correlationId, type: messageType});
+
+      complete('caught', message.content.message, {correlationId});
     }
 
     function onApiMessage(routingKey, message) {
-      const messageType = message.properties.type;
+      const {type: messageType, correlationId} = message.properties;
       switch (messageType) {
         case 'message':
         case 'signal': {
-          return complete('got signal with', message.content.message);
+          return complete('got signal with', message.content.message, {correlationId});
         }
         case 'discard': {
           completed = true;
           stop();
-          return broker.publish('execution', 'execute.discard', {...messageContent});
+          return broker.publish('execution', 'execute.discard', {...messageContent}, {correlationId});
         }
         case 'stop': {
           return stop();
@@ -69,7 +74,7 @@ export default function MessageEventDefinition(activity, eventDefinition) {
       }
     }
 
-    function complete(verb, output) {
+    function complete(verb, output, options) {
       completed = true;
 
       stop();
@@ -82,13 +87,14 @@ export default function MessageEventDefinition(activity, eventDefinition) {
         parent: shiftParent(executeMessage.content.parent),
       }, {type: 'catch'});
 
-      return broker.publish('execution', 'execute.completed', {...messageContent, output, state: 'catch'});
+      return broker.publish('execution', 'execute.completed', {...messageContent, output, state: 'catch'}, options);
     }
 
     function stop() {
       broker.cancel(`_onmessage-${executionId}`);
       broker.cancel(`_api-${executionId}`);
       broker.cancel(`_api-parent-${executionId}`);
+      broker.cancel(`_api-delegated-${executionId}`);
       broker.purgeQueue(messageQueueName);
     }
   }